+++ src/mock/task.js
@@ -1,28 +1,8 @@
 import {DescriptionTasks, DefaultRepeatingDays, Tags, Colors} from '../const';
+import {getRandomDate, getRandomElement, getRandomBoolean} from '../utils';
 
 const generateRepeatingDays = () => {
   return Object.assign({}, DefaultRepeatingDays, {'mo': getRandomBoolean()}, {'we': getRandomBoolean()}, {'fr': getRandomBoolean()});
-};
-
-const getRandomBetween = (min, max) => {
-  return min + Math.floor(Math.random() * (max - min));
-};
-
-const getRandomElement = (array) => {
-  const indexRandom = getRandomBetween(0, array.length);
-
-  return array[indexRandom];
-};
-
-const getRandomBoolean = () => Math.random() > 0.5;
-
-const getRandomDate = () => {
-  const targetDate = new Date();
-  const diffValue = getRandomBetween(0, 15) - 7;
-
-  targetDate.setDate(targetDate.getDate() + diffValue);
-
-  return targetDate;
 };
 
 const generateTags = (tags) => {
